+++ dart/common/detail/Factory-impl.hpp
@@ -178,6 +178,20 @@ HeldT Factory<KeyT, BaseT, HeldT, Args...>::create(
 
 //==============================================================================
 template <typename KeyT,
+         typename BaseT,
+         typename HeldT,
+         typename... Args>
+std::unordered_set<KeyT> Factory<KeyT, BaseT, HeldT, Args...>::getKeys() const
+{
+  std::unordered_set<KeyT> keys;
+  for(const auto& entry : mCreatorMap)
+    keys.insert(entry.first);
+
+  return keys;
+}
+
+//==============================================================================
+template <typename KeyT,
           typename BaseT,
           typename HeldT,
           typename... Args>