+++ dart/io/urdf/DartLoader.hpp
@@ -139,6 +139,11 @@ private:
       const common::Uri& _baseUri,
       const common::ResourceRetrieverPtr& _resourceRetriever);
 
+    static bool addMimicJointsRecursive(
+      const urdf::ModelInterface* model,
+      dynamics::SkeletonPtr _skel,
+      const urdf::Link* _lk);
+
     template <class VisualOrCollision>
     static dynamics::ShapePtr createShape(const VisualOrCollision* _vizOrCol,
       const common::Uri& _baseUri,